♻️ refactor: 상수값 파일명 및 위치 변경
